+++ storage/remote/otlptranslator/prometheusremotewrite/helper.go
@@ -118,7 +118,7 @@ var seps = []byte{'\xff'}
 // if logOnOverwrite is true, the overwrite is logged. Resulting label names are sanitized.
 // If settings.PromoteResourceAttributes is not empty, it's a set of resource attributes that should be promoted to labels.
 func createAttributes(resource pcommon.Resource, attributes pcommon.Map, scope scope, settings Settings,
-	ignoreAttrs []string, logOnOverwrite bool, metadata prompb.MetricMetadata, extras ...string,
+	ignoreAttrs []string, logOnOverwrite bool, metadata prompb.MetricMetadata, temporality pmetric.AggregationTemporality, extras ...string,
 ) ([]prompb.Label, error) {
 	resourceAttrs := resource.Attributes()
 	serviceName, haveServiceName := resourceAttrs.Get(conventions.AttributeServiceName)
@@ -144,6 +144,9 @@ func createAttributes(resource pcommon.Resource, attributes pcommon.Map, scope s
 	}
 	if settings.EnableTypeAndUnitLabels {
 		maxLabelCount += 2
+	}
+	if settings.AllowDeltaTemporality && (temporality == pmetric.AggregationTemporalityCumulative || temporality == pmetric.AggregationTemporalityDelta) {
+		maxLabelCount++
 	}
 
 	// Ensure attributes are sorted by key for consistent merging of keys which
@@ -205,11 +208,24 @@ func createAttributes(resource pcommon.Resource, attributes pcommon.Map, scope s
 
 	if settings.EnableTypeAndUnitLabels {
 		unitNamer := otlptranslator.UnitNamer{UTF8Allowed: settings.AllowUTF8}
+
 		if metadata.Type != prompb.MetricMetadata_UNKNOWN {
 			l["__type__"] = strings.ToLower(metadata.Type.String())
 		}
 		if metadata.Unit != "" {
 			l["__unit__"] = unitNamer.Build(metadata.Unit)
+		}
+	}
+	if settings.AllowDeltaTemporality {
+		switch temporality {
+		case pmetric.AggregationTemporalityCumulative:
+			l["__temporality__"] = "cumulative"
+		case pmetric.AggregationTemporalityDelta:
+			l["__temporality__"] = "delta"
+		case pmetric.AggregationTemporalityUnspecified:
+			// For metric types without temporality (Gauge, Summary), we don't add a temporality label.
+		default:
+			return nil, fmt.Errorf("unknown aggregation temporality: %d", temporality)
 		}
 	}
 
@@ -286,7 +302,7 @@ func aggregationTemporality(metric pmetric.Metric) (pmetric.AggregationTemporali
 // However, work is under way to resolve this shortcoming through a feature called native histograms custom buckets:
 // https://github.com/prometheus/prometheus/issues/13485.
 func (c *PrometheusConverter) addHistogramDataPoints(ctx context.Context, dataPoints pmetric.HistogramDataPointSlice,
-	resource pcommon.Resource, settings Settings, metadata prompb.MetricMetadata, scope scope,
+	resource pcommon.Resource, settings Settings, metadata prompb.MetricMetadata, scope scope, temporality pmetric.AggregationTemporality,
 ) error {
 	for x := 0; x < dataPoints.Len(); x++ {
 		if err := c.everyN.checkContext(ctx); err != nil {
@@ -295,7 +311,7 @@ func (c *PrometheusConverter) addHistogramDataPoints(ctx context.Context, dataPo
 
 		pt := dataPoints.At(x)
 		timestamp := convertTimeStamp(pt.Timestamp())
-		baseLabels, err := createAttributes(resource, pt.Attributes(), scope, settings, nil, false, metadata)
+		baseLabels, err := createAttributes(resource, pt.Attributes(), scope, settings, nil, false, metadata, temporality)
 		if err != nil {
 			return err
 		}
@@ -502,7 +518,7 @@ func (c *PrometheusConverter) addSummaryDataPoints(ctx context.Context, dataPoin
 
 		pt := dataPoints.At(x)
 		timestamp := convertTimeStamp(pt.Timestamp())
-		baseLabels, err := createAttributes(resource, pt.Attributes(), scope, settings, nil, false, metadata)
+		baseLabels, err := createAttributes(resource, pt.Attributes(), scope, settings, nil, false, metadata, pmetric.AggregationTemporalityUnspecified)
 		if err != nil {
 			return err
 		}
@@ -650,7 +666,7 @@ func addResourceTargetInfo(resource pcommon.Resource, settings Settings, earlies
 		// Do not pass identifying attributes as ignoreAttrs below.
 		identifyingAttrs = nil
 	}
-	labels, err := createAttributes(resource, attributes, scope{}, settings, identifyingAttrs, false, prompb.MetricMetadata{}, model.MetricNameLabel, name)
+	labels, err := createAttributes(resource, attributes, scope{}, settings, identifyingAttrs, false, prompb.MetricMetadata{}, pmetric.AggregationTemporalityUnspecified, model.MetricNameLabel, name)
 	if err != nil {
 		return err
 	}