+++ storage/remote/otlptranslator/prometheusremotewrite/histograms.go
@@ -60,6 +60,7 @@ func (c *PrometheusConverter) addExponentialHistogramDataPoints(ctx context.Cont
 			nil,
 			true,
 			metadata,
+			temporality,
 			model.MetricNameLabel,
 			metadata.MetricFamilyName,
 		)
@@ -281,6 +282,7 @@ func (c *PrometheusConverter) addCustomBucketsHistogramDataPoints(ctx context.Co
 			nil,
 			true,
 			metadata,
+			temporality,
 			model.MetricNameLabel,
 			metadata.MetricFamilyName,
 		)